nrod_consumer/requirements.txt
requests
boto3
stomp.py==4.1.22